cryptoprices/cloudfunction/requirements.txt
google-cloud>=0.34.0
google-cloud-secret-manager>=2.4.0
google-cloud-storage>=1.37.0
requests>=2.22.0